Austin-develop (#7)
* get second implementation created

* Reorganized files.

* Ready to test GEMMTP.

* Rough prototype of GEMM tensor product.

* More progress on GEMM tensor product.

* Updated CmakeLists.txt to link against cuBLASLt.

* cuBLASTLt works for equal dimensions and batch size 1. Need to tune / debug.

* Tested all three cases for batch size 1.

* GEMM tensor product using cuBLAS is working.

* Made a GPU timer class.

* Minor changes.

* Ready to test GEMMTP.

* Ready to test GEMMTP.

* ready to do new work

* remove my addition

* remove wip

* fix deleted }

* 'fix title'

* fix title

* adding exec_tesnor_product for direct GPU support

* add get_gpu_node.sh to gitignore

* put account back to how it should be

* remove whitespace

* remove space

* fix exec_tesnor_product issue

* support multiplicty for thread tp

* requested changes, mult, type, even now ints

---------

+++ src/kernel/util/representation.hpp
@@ -35,15 +35,15 @@ public:
         return irreps.size();
     }
 
-    size_t mult(int irrep_id) {
+    int mult(int irrep_id) {
         return get<0>(irreps[irrep_id]);
     }
 
-    size_t type(int irrep_id) {
+    int type(int irrep_id) {
         return get<1>(irreps[irrep_id]);
     }
 
-    size_t even(int irrep_id) {
+    int even(int irrep_id) {
         return get<2>(irreps[irrep_id]);
     }
